Integrate the VCL code closer in the cache process:  The passed argument will be the session structure.
Add the pool of worker threads (cache_pool).  With a unitary nature
of the VCL code, the HTTP parsing can be postponed to the worker thread.

This actually helps us with CPU cache locality as it will reduce the
amount of memory allocated on one CPU but freed on another.




git-svn-id: http://varnish-cache.org/svn/trunk@84 d4fa192b-c00b-0410-8231-f00ffab90ce4