Put more meat on the stevedore (storage backend) interface:
Pull the struct definition into _stevedore.h and include this from
cache.h and mgt.h, they both need to be able to see it.

Add the stevedore pointer as an argument to the stevedore alloc function
so multiple stevedores is possible later on.

Add the stevedore pointer to the storage object, so freeing it again is
possible.

Add -s argument processing to select a given stevedore, call it's ->init
method and pass the stevedore in the heritage.

In the cache process pick stevedore out from heritage, call its open method.




git-svn-id: http://varnish-cache.org/svn/trunk@170 d4fa192b-c00b-0410-8231-f00ffab90ce4