VCL compiler: 	add two sbufs for "init" and "fini" actions.
VCL ACLs:  Change syntax and implementation as follows.

ACL Syntax now works the following way:

	acl $name {
		! ( "myhost.com" ) ;
		"10.0.0.1" /8 ;
	}

The '!' means not.  If the address matches the rest of the rule
the address does NOT match the acl and the search terminates here.

Enclosing the string in paranthesis means that the rule will be ignored
if the string cannot be converted to an address (with getaddrinfo).

When a string can not be looked up, and is not enclosed in a
paranthesis, a positive rule (ie: without !) will not match and a
negative rule (with !) will match.

A mask can always be supplied, no matter the style of the string
given, so it is possible to do things like:

	{ "fw.ourcompany.dom" / 24 }

Which means "any host on the same /24 subnet as fw.ourcompany.dom".


Unfortunately getaddrinfo() does not return a TTL for the results,
in the future we may want to use some kind of timeout to refresh
the lookups.




git-svn-id: http://varnish-cache.org/svn/trunk@556 d4fa192b-c00b-0410-8231-f00ffab90ce4