Add per address, per session and total statistics.
We (will) collect data in unlocked per workerthread accumulators
and whenever the workerthread leaves the session, we charge the
bill to the srcaddr (issuing a StatAddr shmrecord), to the session
and to the global counters in the stats struct.

When sessions die we issue a StatSess shmrecord.

StatAddr and StatSess has the same format:
	address
	port (always zero for StatAddr)
	duration (seconds)
	#sessions
	#requests
	#pipe
	#pass
	#fetch
	#hdrbytes
	#bodybytes



git-svn-id: http://varnish-cache.org/svn/trunk@564 d4fa192b-c00b-0410-8231-f00ffab90ce4