From the "What has the world come to ?" department:
It used to be that sscanf(3) would walk along the proffered string
and do what the format string was told, and as long as the format
terminated before the input string, the NUL termination was not
relevant.

Sometime in the last 20 years, sscanf started doing a strlen on the
passed argument changing this behaviour.

Give sscanf the NUL it wants.



git-svn-id: http://varnish-cache.org/svn/trunk@1705 d4fa192b-c00b-0410-8231-f00ffab90ce4