Rewrite the req.hash implmentation:
Instead of assembling the entire hash-string in the workspace, use
a scatter gather approach, hinted by the VCL compiler.

This eliminates the workspace reservation which prevented regsub() from
working in vcl_hash, and reduces the size of the necessary workspace a
fair bit as well, at the cost of a little bit of complexity in the
hash implmentations.

Closes ticket 137 and possibly 141



git-svn-id: http://varnish-cache.org/svn/trunk@1805 d4fa192b-c00b-0410-8231-f00ffab90ce4