Add esi:comment support.  "Support" is defined as: Silently ignore, as opposed to unhandled esi: elments which we complain about in varnishlog.
Thus, if you want to get a short comment into the shmlog, the
easiest way is to do something like <esi:say a="Hi Mom"> which will
result in a shmlog record:
   11 ESI_xmlerror c at 25: ESI 1.0 unimplemented element "<esi:say Hi Mom>"
But the length of the message is truncated to avoid dumping the entire
source document into the shmlog.

Snip out any unknown esi: element.

While the ESI 1.0 specification doesn't address this directly, my
impression from the document is that they should never leak through
to the client.



git-svn-id: http://varnish-cache.org/svn/trunk@2282 d4fa192b-c00b-0410-8231-f00ffab90ce4