I am not sure if this is a/the race some users are seeing, or if it even can have any effect, but this will close it at a cost of one extra kevent(2) every 100ms timer tick.
The (perceived) problem is that we have pending kqueue changes we
have not yet told the kernel about, then close a number of expired
FD's which might be instantly be recycled by the accept(2) over in
the other thread before we tell the kernel about the pending changes.

In that case, the kernel has no way of knowing that our changes
referred to the previous instance of the fd and not the new one.

The solution is to push the changes to the kernel before servicing
the timer.



git-svn-id: http://varnish-cache.org/svn/trunk@2401 d4fa192b-c00b-0410-8231-f00ffab90ce4