Rework the manager/cacher cli relationship, so that the manager does not have to grope around in the cacher's data structures.
Whenever the manager gets a CLI command it does not understand, it
will pass it to the cacher process, if it is running.

This complicated "help" a little bit, because we want to show the
combined commands of the manager and cacher.

Since we're dealing with help anyway, hide undocumented debug
commands from it.



git-svn-id: http://varnish-cache.org/svn/trunk@2595 d4fa192b-c00b-0410-8231-f00ffab90ce4