Slight backtracking:  Store the specified socket name in the listen socket structure, instead of whatever we actually bound to.  It is more informative and intuitive.
Report the listen socket spec in the SessionOpen shm record.

Close listen sockets as soon as we have forked the child, this eliminated
a fd-leak when a socket was eliminated from listen_address while the child
ran.

Fix various potential problems related to not being able to bind to one
or more listen addresses.



git-svn-id: http://varnish-cache.org/svn/trunk@2659 d4fa192b-c00b-0410-8231-f00ffab90ce4