Add the ability to purge on random request or object headers.
For instance:

	purge req.http.host ~ www.foo.com && req.url ~ "\.png$"
	purge obj.http.set-cookie ~ USER=383839

Now, why would you want purge on request headers and not object headers ?

Simple, some information the object does not have, the Host: header is
a good example.

Assuming that the Host: header is part of the hash we use to lookup
an object (as is the default), we can avoid copying that field into
the object (saving memory: O(nObjects)) by using the request value
to purge against.




git-svn-id: http://varnish-cache.org/svn/trunk@3519 d4fa192b-c00b-0410-8231-f00ffab90ce4