Give objects a back-pointer to their entry in the segment object index.
Have the objcore point to the smp_object until fixup, so that we catch
any attempt to muck with it before that and so that we hve the address
available in Fixup.

Since we don't have the silos coordinates at fixup time, make the
smp_object point to the object.

Add callbacks for updating BAN and TTL information in the segment object
index.



git-svn-id: http://varnish-cache.org/svn/trunk@3914 d4fa192b-c00b-0410-8231-f00ffab90ce4