Save and restore bans from persistent storage.
We append bans in all silos all the time.

(Still missing: periodic rewrite of list to get rid of "gone" bans)

When we open the silos, we read in all the bans in the silo before
kicking of the silo thread that reads on the object indicies, this
way we are sure to have the bans referenced from the objects.

Add a testcase to see that duplicate elimination (from multiple silos) work.



git-svn-id: http://varnish-cache.org/svn/trunk@3939 d4fa192b-c00b-0410-8231-f00ffab90ce4