Exploit the new and cleaner director interface and add a "hash" director by sharing most of the code with the "random" director.
The Hash director will use the hash output from vcl_hash{} to pick a
backend from the configured set.

If this backend is unhealthy, it pick another backend, using only the
healthy subset of backends, but as soon as the "canonical" backend becomes
healthy it will revert to using that.

This means that one backend going unhealthy on "rehashes" the requests for
that backend, hashing for healthy backends are not affected.





git-svn-id: http://varnish-cache.org/svn/trunk@4419 d4fa192b-c00b-0410-8231-f00ffab90ce4