Add the master and child pid's to the SHMFILE.
If the master pid is active when we start, we issue a message about
this, (with a hint about -n) and exit.

If the master pid is not active, but the child pid is, we issue a
message about it presumably being busy dying, dump the SHMFILE and
create a new one.  (This only saves our bacon, if the dying
process manages to close the listening sockets before we need them).

This should end any confusion that might arise from accidentally
running multiple varnishes at the same time.

Fixes #620



git-svn-id: http://varnish-cache.org/svn/trunk@4520 d4fa192b-c00b-0410-8231-f00ffab90ce4