Random director tries all backends before giving up
Summary:
The current random director gives up when it can't get a FD to the backend
it wants retries times in a row.  Rather than give up and return NULL,
which is guaranteed to cause a vcl_error, as a last ditch effort we try
all other healthy backends until we get one that works.  This is mostly
useful in the between time after a backend server dies and before the
health check fails enough to mark a backend unhealthy.

Backwards Compatibility =  Not strictly backwards compatible.  In cases
when the old code would of fallen through to vcl_error this will give a
shot at getting a good result.

Performance = In the worse case, this will add extra calls for getting a
FD, but only for situations that vcl_error

Test Plan: New varnish unittest.  It fails in the old code and works in
this new code.